maintain the current mileage in the shoe sheet.
so it's that total_miles value that we want to push back into the
sheet, but where...

The YTD column (F in spreadsheet, iloc=5) should hold total to now
 - ShoeManager already has updateTotals which will write back from
   DataFrame to the sheet.
The hashes column, needs just the total for this year.

It looks like the ShoeMileYears is the right place to put the updates
of the DataFrame - leveraging that total_miles line which has all the
numbers we want.
 - ShoeMileYears.save() would write back an updated hash
 - create a ShoeMileYear.total(yearlyShoeSum, year) which:
     - pushes the year total into the YTD col (how)
     - calculates the total using self.residual(year)
     - pushes total into hash for year
     - returns total